requirements.txt
openai>=1.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
python-dotenv>=1.0.0
Pillow>=10.0.0
